d_a_npc_aj1: wait_action1 90.2->100.0 (explicit cases 1-3 not default — retail dispatch pivots 4/0/9), talk_1 u8->int return (retail callers store r3 w/o clrlwi)

+++ include/d/actor/d_a_npc_aj1.h
@@ -94,7 +94,7 @@ public:
     s32 FARwai();
     s32 call_1();
     s32 wait_1();
-    u8 talk_1();
+    int talk_1();
     int wait_action1(void*);
     int wait_action2(void*);
     u8 demo();
